Add dto to entity mapping

+++ retail-be/src/main/java/com/utm/retail_be/csv/dto/DateDetailsDTO.java
@@ -11,6 +11,6 @@ public class DateDetailsDTO {
     private int id;
     private LocalDate dataBonData;
     private int sarbatoare;
-    private int ziua;
+    private short ziua;
 
 }
